+++ src/unitycatalog/ai/utils/config.py
@@ -1,0 +1,19 @@
+UC_LIST_FUNCTIONS_MAX_RESULTS = "100"
+
+# JSON schema definition: https://json-schema.org/understanding-json-schema/reference/type
+JSON_SCHEMA_TYPE = {
+    "ARRAY",
+    "BOOLEAN",
+    "BYTE",
+    "CHAR",
+    "DOUBLE",
+    "FLOAT",
+    "INT",
+    "LONG",
+    "MAP",
+    "NULL",
+    "SHORT",
+    "STRING",
+    "STRUCT",
+    "TABLE_TYPE",
+}
