+++ src/unitycatalog/ai/utils/pydantic_utils.py
@@ -1,0 +1,23 @@
+from dataclasses import dataclass
+from pydantic import BaseModel
+from typing import Optional, Type, Any
+
+
+@dataclass
+class PydanticType:
+    pydantic_type: Type
+    strict: bool = False
+
+
+@dataclass
+class PydanticField:
+    pydantic_type: Type
+    description: Optional[str] = None
+    default: Optional[Any] = None
+    strict: bool = False
+
+
+@dataclass
+class PydanticFunctionInputParams:
+    pydantic_model: Type[BaseModel]
+    strict: bool = False
