+++ tests/ai/test_utils.py
@@ -3,14 +3,16 @@ import datetime
 import decimal
 from typing import Union, List, Optional, Tuple, Dict
 from pydantic import BaseModel
-from unitycatalog.ai.utils import (
-    column_type_to_python_type,
-    convert_timedelta_to_interval_str,
+from unitycatalog.ai.utils.function_processing_utils import (
     validate_full_function_name,
     uc_type_json_to_pydantic_type,
     generate_function_input_params_schema,
-    is_base64_encoded,
 )
+from unitycatalog.ai.utils.type_utils import (
+    column_type_to_python_type,
+    convert_timedelta_to_interval_str,
+)
+from unitycatalog.ai.utils.validation_utils import is_base64_encoded
 from databricks.sdk.service.catalog import (
     FunctionParameterInfo,
     FunctionInfo,
